day 7

+++ aoc/2024/_proj/string_vector.h
@@ -135,7 +135,7 @@ cIntVector tStringVector<StoredType>::toIntVector() const
         {
             if constexpr (std::is_same_v<StoredType, std::string>)
             {
-                intVector[idx] = std::stoi(s);
+                intVector[idx] = std::stoll(s);
             }
             else
             {
